+++ crates/opsml_sql/src/sqlite/client.rs
@@ -1,51 +1,51 @@
-use crate::base::SqlClient;
-
-use crate::error::SqlError;
-use crate::schemas::schema::{
-    ArtifactSqlRecord, AuditCardRecord, CardResults, CardSummary, DataCardRecord,
-    ExperimentCardRecord, HardwareMetricsRecord, MetricRecord, ModelCardRecord, ParameterRecord,
-    PromptCardRecord, QueryStats, ServerCard, ServiceCardRecord, SqlSpaceRecord, User,
-    VersionResult, VersionSummary,
+use crate::{
+    error::SqlError,
+    sqlite::sql::{
+        artifact::ArtifactLogicSqliteClient, audit::SpaceLogicSqliteClient,
+        card::CardLogicSqliteClient, experiment::ExperimentLogicSqliteClient,
+        space::AuditLogicSqliteClient, user::UserLogicSqliteClient,
+    },
 };
-use crate::sqlite::helper::SqliteQueryHelper;
-use async_trait::async_trait;
-use opsml_semver::VersionValidator;
 use opsml_settings::config::DatabaseSettings;
-use opsml_types::contracts::{
-    ArtifactKey, ArtifactQueryArgs, ArtifactRecord, AuditEvent, SpaceNameEvent, SpaceRecord,
-    SpaceStats,
-};
-use opsml_types::{cards::CardTable, contracts::CardQueryArgs, RegistryType};
-use semver::Version;
 use sqlx::{sqlite::SqlitePoolOptions, Pool, Sqlite};
-use tracing::{debug, error, instrument};
+use std::path::Path;
+use tracing::{debug, instrument};
 
 #[derive(Debug, Clone)]
 pub struct SqliteClient {
     pub pool: Pool<Sqlite>,
+
+    // The SqliteClient is a parent struct that holds instances of various logic clients.
+    // Each logic client is responsible for a specific domain (e.g., cards, experiments, etc).
+    pub card: CardLogicSqliteClient,
+    pub exp: ExperimentLogicSqliteClient,
+    pub artifact: ArtifactLogicSqliteClient,
+    pub user: UserLogicSqliteClient,
+    pub space: SpaceLogicSqliteClient,
+    pub audit: AuditLogicSqliteClient,
 }
 
-#[async_trait]
-impl SqlClient for SqliteClient {
-    async fn new(settings: &DatabaseSettings) -> Result<Self, SqlError> {
-        // if the connection uri is not in memory, create the file
+impl SqliteClient {
+    pub async fn new(settings: &DatabaseSettings) -> Result<SqliteClient, SqlError> {
+        // Create SQLite file if it doesn't exist and not in-memory
         if !settings.connection_uri.contains(":memory:") {
-            // strip "sqlite://" from the connection uri
             let uri = settings.connection_uri.replace("sqlite://", "");
-            let path = std::path::Path::new(&uri);
+            let path = Path::new(&uri);
 
-            // check if the file exists
             if !path.exists() {
-                debug!("SQLite file does not exist, creating file");
-                // Ensure the parent directory exists
+                debug!("SQLite file does not exist, creating file at: {}", uri);
+
+                // Ensure parent directory exists
                 if let Some(parent) = path.parent() {
                     if !parent.exists() {
-                        std::fs::create_dir_all(parent)?;
+                        std::fs::create_dir_all(parent)
+                            .map_err(|e| SqlError::ConnectionError(sqlx::Error::Io(e)))?;
                     }
                 }
 
-                // create the file
-                std::fs::File::create(&uri)?;
+                // Create the file
+                std::fs::File::create(&uri)
+                    .map_err(|e| SqlError::ConnectionError(sqlx::Error::Io(e)))?;
             }
         }
 
@@ -55,1203 +55,63 @@ impl SqlClient for SqliteClient {
             .await
             .map_err(SqlError::ConnectionError)?;
 
-        let client = Self { pool };
+        let client = SqliteClient {
+            card: CardLogicSqliteClient::new(&pool),
+            exp: ExperimentLogicSqliteClient::new(&pool),
+            artifact: ArtifactLogicSqliteClient::new(&pool),
+            user: UserLogicSqliteClient::new(&pool),
+            space: SpaceLogicSqliteClient::new(&pool),
+            audit: AuditLogicSqliteClient::new(&pool),
+            pool,
+        };
 
-        // run migrations
+        // Run migrations
         client.run_migrations().await?;
 
+        debug!("SQLite client initialized successfully");
         Ok(client)
     }
-    async fn run_migrations(&self) -> Result<(), SqlError> {
-        debug!("Running migrations");
+
+    #[instrument(skip(self))]
+    pub async fn run_migrations(&self) -> Result<(), SqlError> {
+        debug!("Running SQLite migrations");
         sqlx::migrate!("src/sqlite/migrations")
             .run(&self.pool)
             .await
             .map_err(SqlError::MigrationError)?;
+        debug!("SQLite migrations completed successfully");
         Ok(())
     }
+}
 
-    /// Check if uid exists in the database for a table
-    ///
-    /// # Arguments
-    ///
-    /// * `table` - The table to query
-    /// * `uid` - The uid to check
-    ///
-    /// # Returns
-    ///
-    /// * `bool` - True if the uid exists, false otherwise
-    async fn check_uid_exists(&self, uid: &str, table: &CardTable) -> Result<bool, SqlError> {
-        let query = SqliteQueryHelper::get_uid_query(table);
-        let exists: Option<String> = sqlx::query_scalar(&query)
-            .bind(uid)
-            .fetch_optional(&self.pool)
-            .await?;
-
-        Ok(exists.is_some())
-    }
-
-    /// Primary query for retrieving versions from the database. Mainly used to get most recent version when determining version increment
-    ///
-    /// # Arguments
-    ///
-    /// * `table` - The table to query
-    /// * `name` - The name of the card
-    /// * `space` - The space of the card
-    /// * `version` - The version of the card
-    ///
-    /// # Returns
-    ///
-    /// * `Vec<String>` - A vector of strings representing the sorted (desc) versions of the card
-    #[instrument(skip_all)]
-    async fn get_versions(
-        &self,
-        table: &CardTable,
-        space: &str,
-        name: &str,
-        version: Option<String>,
-    ) -> Result<Vec<String>, SqlError> {
-        // if version is None, get the latest version
-
-        let query = SqliteQueryHelper::get_versions_query(table, version)?;
-
-        let cards: Vec<VersionResult> = sqlx::query_as(&query)
-            .bind(name)
-            .bind(space)
-            .fetch_all(&self.pool)
-            .await?;
-
-        let versions = cards
-            .iter()
-            .map(|c| c.to_version())
-            .collect::<Result<Vec<Version>, SqlError>>()?;
-
-        // sort semvers
-        Ok(
-            VersionValidator::sort_semver_versions(versions, true).inspect_err(|e| {
-                error!("{}", e);
-            })?,
-        )
-    }
-
-    /// Query cards based on the query arguments
-    ///
-    /// # Arguments
-    ///
-    /// * `table` - The table to query
-    /// * `query_args` - The query arguments
-    ///
-    /// # Returns
-    ///
-    /// * `CardResults` - The results of the query
-    async fn query_cards(
-        &self,
-        table: &CardTable,
-        query_args: &CardQueryArgs,
-    ) -> Result<CardResults, SqlError> {
-        let query = SqliteQueryHelper::get_query_cards_query(table, query_args)?;
-
-        match table {
-            CardTable::Data => {
-                let card: Vec<DataCardRecord> = sqlx::query_as(&query)
-                    .bind(query_args.uid.as_ref())
-                    .bind(query_args.space.as_ref())
-                    .bind(query_args.name.as_ref())
-                    .bind(query_args.max_date.as_ref())
-                    .bind(query_args.limit.unwrap_or(50))
-                    .fetch_all(&self.pool)
-                    .await?;
-
-                return Ok(CardResults::Data(card));
-            }
-            CardTable::Model => {
-                let card: Vec<ModelCardRecord> = sqlx::query_as(&query)
-                    .bind(query_args.uid.as_ref())
-                    .bind(query_args.space.as_ref())
-                    .bind(query_args.name.as_ref())
-                    .bind(query_args.max_date.as_ref())
-                    .bind(query_args.limit.unwrap_or(50))
-                    .fetch_all(&self.pool)
-                    .await?;
-
-                return Ok(CardResults::Model(card));
-            }
-            CardTable::Experiment => {
-                let card: Vec<ExperimentCardRecord> = sqlx::query_as(&query)
-                    .bind(query_args.uid.as_ref())
-                    .bind(query_args.space.as_ref())
-                    .bind(query_args.name.as_ref())
-                    .bind(query_args.max_date.as_ref())
-                    .bind(query_args.limit.unwrap_or(50))
-                    .fetch_all(&self.pool)
-                    .await?;
-
-                return Ok(CardResults::Experiment(card));
-            }
-
-            CardTable::Audit => {
-                let card: Vec<AuditCardRecord> = sqlx::query_as(&query)
-                    .bind(query_args.uid.as_ref())
-                    .bind(query_args.space.as_ref())
-                    .bind(query_args.name.as_ref())
-                    .bind(query_args.max_date.as_ref())
-                    .bind(query_args.limit.unwrap_or(50))
-                    .fetch_all(&self.pool)
-                    .await?;
-
-                return Ok(CardResults::Audit(card));
-            }
-
-            CardTable::Prompt => {
-                let card: Vec<PromptCardRecord> = sqlx::query_as(&query)
-                    .bind(query_args.uid.as_ref())
-                    .bind(query_args.space.as_ref())
-                    .bind(query_args.name.as_ref())
-                    .bind(query_args.max_date.as_ref())
-                    .bind(query_args.limit.unwrap_or(50))
-                    .fetch_all(&self.pool)
-                    .await?;
-
-                return Ok(CardResults::Prompt(card));
-            }
-
-            CardTable::Service => {
-                let card: Vec<ServiceCardRecord> = sqlx::query_as(&query)
-                    .bind(query_args.uid.as_ref())
-                    .bind(query_args.space.as_ref())
-                    .bind(query_args.name.as_ref())
-                    .bind(query_args.max_date.as_ref())
-                    .bind(query_args.limit.unwrap_or(50))
-                    .fetch_all(&self.pool)
-                    .await?;
-
-                return Ok(CardResults::Service(card));
-            }
-
-            _ => {
-                return Err(SqlError::InvalidTableName);
-            }
-        }
-    }
-
-    async fn insert_card(&self, table: &CardTable, card: &ServerCard) -> Result<(), SqlError> {
-        match table {
-            CardTable::Data => match card {
-                ServerCard::Data(record) => {
-                    let query = SqliteQueryHelper::get_datacard_insert_query();
-                    sqlx::query(&query)
-                        .bind(&record.uid)
-                        .bind(&record.app_env)
-                        .bind(&record.name)
-                        .bind(&record.space)
-                        .bind(record.major)
-                        .bind(record.minor)
-                        .bind(record.patch)
-                        .bind(&record.version)
-                        .bind(&record.data_type)
-                        .bind(&record.interface_type)
-                        .bind(&record.tags)
-                        .bind(&record.experimentcard_uid)
-                        .bind(&record.auditcard_uid)
-                        .bind(&record.pre_tag)
-                        .bind(&record.build_tag)
-                        .bind(&record.username)
-                        .bind(&record.opsml_version)
-                        .execute(&self.pool)
-                        .await?;
-                    Ok(())
-                }
-                _ => {
-                    return Err(SqlError::InvalidCardType);
-                }
-            },
-            CardTable::Model => match card {
-                ServerCard::Model(record) => {
-                    let query = SqliteQueryHelper::get_modelcard_insert_query();
-                    sqlx::query(&query)
-                        .bind(&record.uid)
-                        .bind(&record.app_env)
-                        .bind(&record.name)
-                        .bind(&record.space)
-                        .bind(record.major)
-                        .bind(record.minor)
-                        .bind(record.patch)
-                        .bind(&record.version)
-                        .bind(&record.datacard_uid)
-                        .bind(&record.data_type)
-                        .bind(&record.model_type)
-                        .bind(&record.interface_type)
-                        .bind(&record.task_type)
-                        .bind(&record.tags)
-                        .bind(&record.experimentcard_uid)
-                        .bind(&record.auditcard_uid)
-                        .bind(&record.pre_tag)
-                        .bind(&record.build_tag)
-                        .bind(&record.username)
-                        .bind(&record.opsml_version)
-                        .execute(&self.pool)
-                        .await?;
-                    Ok(())
-                }
-                _ => {
-                    return Err(SqlError::InvalidCardType);
-                }
-            },
-            CardTable::Experiment => match card {
-                ServerCard::Experiment(record) => {
-                    let query = SqliteQueryHelper::get_experimentcard_insert_query();
-                    sqlx::query(&query)
-                        .bind(&record.uid)
-                        .bind(&record.app_env)
-                        .bind(&record.name)
-                        .bind(&record.space)
-                        .bind(record.major)
-                        .bind(record.minor)
-                        .bind(record.patch)
-                        .bind(&record.version)
-                        .bind(&record.tags)
-                        .bind(&record.datacard_uids)
-                        .bind(&record.modelcard_uids)
-                        .bind(&record.promptcard_uids)
-                        .bind(&record.service_card_uids)
-                        .bind(&record.experimentcard_uids)
-                        .bind(&record.pre_tag)
-                        .bind(&record.build_tag)
-                        .bind(&record.username)
-                        .bind(&record.opsml_version)
-                        .execute(&self.pool)
-                        .await?;
-                    Ok(())
-                }
-                _ => {
-                    return Err(SqlError::InvalidCardType);
-                }
-            },
-            CardTable::Audit => match card {
-                ServerCard::Audit(record) => {
-                    let query = SqliteQueryHelper::get_auditcard_insert_query();
-                    sqlx::query(&query)
-                        .bind(&record.uid)
-                        .bind(&record.app_env)
-                        .bind(&record.name)
-                        .bind(&record.space)
-                        .bind(record.major)
-                        .bind(record.minor)
-                        .bind(record.patch)
-                        .bind(&record.version)
-                        .bind(&record.tags)
-                        .bind(record.approved)
-                        .bind(&record.datacard_uids)
-                        .bind(&record.modelcard_uids)
-                        .bind(&record.experimentcard_uids)
-                        .bind(&record.pre_tag)
-                        .bind(&record.build_tag)
-                        .bind(&record.username)
-                        .bind(&record.opsml_version)
-                        .execute(&self.pool)
-                        .await?;
-                    Ok(())
-                }
-                _ => {
-                    return Err(SqlError::InvalidCardType);
-                }
-            },
-
-            CardTable::Prompt => match card {
-                ServerCard::Prompt(record) => {
-                    let query = SqliteQueryHelper::get_promptcard_insert_query();
-                    sqlx::query(&query)
-                        .bind(&record.uid)
-                        .bind(&record.app_env)
-                        .bind(&record.name)
-                        .bind(&record.space)
-                        .bind(record.major)
-                        .bind(record.minor)
-                        .bind(record.patch)
-                        .bind(&record.version)
-                        .bind(&record.tags)
-                        .bind(&record.experimentcard_uid)
-                        .bind(&record.auditcard_uid)
-                        .bind(&record.pre_tag)
-                        .bind(&record.build_tag)
-                        .bind(&record.username)
-                        .bind(&record.opsml_version)
-                        .execute(&self.pool)
-                        .await?;
-                    Ok(())
-                }
-
-                _ => {
-                    return Err(SqlError::InvalidCardType);
-                }
-            },
-            CardTable::Service => match card {
-                ServerCard::Service(record) => {
-                    let query = SqliteQueryHelper::get_servicecard_insert_query();
-                    sqlx::query(&query)
-                        .bind(&record.uid)
-                        .bind(&record.app_env)
-                        .bind(&record.name)
-                        .bind(&record.space)
-                        .bind(record.major)
-                        .bind(record.minor)
-                        .bind(record.patch)
-                        .bind(&record.version)
-                        .bind(&record.pre_tag)
-                        .bind(&record.build_tag)
-                        .bind(&record.cards)
-                        .bind(&record.username)
-                        .bind(&record.opsml_version)
-                        .execute(&self.pool)
-                        .await?;
-                    Ok(())
-                }
-                _ => {
-                    return Err(SqlError::InvalidCardType);
-                }
-            },
-
-            _ => {
-                return Err(SqlError::InvalidTableName);
-            }
-        }
-    }
-
-    async fn insert_artifact_record(&self, record: &ArtifactSqlRecord) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_artifact_record_insert_query();
-        sqlx::query(&query)
-            .bind(&record.uid)
-            .bind(record.created_at)
-            .bind(&record.app_env)
-            .bind(&record.space)
-            .bind(&record.name)
-            .bind(record.major)
-            .bind(record.minor)
-            .bind(record.patch)
-            .bind(&record.pre_tag)
-            .bind(&record.build_tag)
-            .bind(&record.version)
-            .bind(&record.media_type)
-            .bind(&record.artifact_type)
-            .execute(&self.pool)
-            .await?;
-        Ok(())
-    }
-
-    async fn query_artifacts(
-        &self,
-        query_args: &ArtifactQueryArgs,
-    ) -> Result<Vec<ArtifactRecord>, SqlError> {
-        let query = SqliteQueryHelper::get_query_artifacts_query(query_args)?;
-        let rows: Vec<ArtifactSqlRecord> = sqlx::query_as(&query)
-            .bind(query_args.uid.as_ref())
-            .bind(query_args.artifact_type.as_ref().map(|a| a.to_string()))
-            .bind(query_args.space.as_ref())
-            .bind(query_args.name.as_ref())
-            .bind(query_args.limit.unwrap_or(50))
-            .fetch_all(&self.pool)
-            .await?;
-
-        Ok(rows
-            .iter()
-            .map(|r| r.to_artifact_record())
-            .collect::<Vec<ArtifactRecord>>())
-    }
-
-    async fn update_card(&self, table: &CardTable, card: &ServerCard) -> Result<(), SqlError> {
-        match table {
-            CardTable::Data => match card {
-                ServerCard::Data(record) => {
-                    let query = SqliteQueryHelper::get_datacard_update_query();
-                    sqlx::query(&query)
-                        .bind(&record.app_env)
-                        .bind(&record.name)
-                        .bind(&record.space)
-                        .bind(record.major)
-                        .bind(record.minor)
-                        .bind(record.patch)
-                        .bind(&record.version)
-                        .bind(&record.data_type)
-                        .bind(&record.interface_type)
-                        .bind(&record.tags)
-                        .bind(&record.experimentcard_uid)
-                        .bind(&record.auditcard_uid)
-                        .bind(&record.pre_tag)
-                        .bind(&record.build_tag)
-                        .bind(&record.username)
-                        .bind(&record.opsml_version)
-                        .bind(&record.uid)
-                        .execute(&self.pool)
-                        .await?;
-                    Ok(())
-                }
-                _ => {
-                    return Err(SqlError::InvalidCardType);
-                }
-            },
-            CardTable::Model => match card {
-                ServerCard::Model(record) => {
-                    let query = SqliteQueryHelper::get_modelcard_update_query();
-                    sqlx::query(&query)
-                        .bind(&record.app_env)
-                        .bind(&record.name)
-                        .bind(&record.space)
-                        .bind(record.major)
-                        .bind(record.minor)
-                        .bind(record.patch)
-                        .bind(&record.version)
-                        .bind(&record.datacard_uid)
-                        .bind(&record.data_type)
-                        .bind(&record.model_type)
-                        .bind(&record.interface_type)
-                        .bind(&record.task_type)
-                        .bind(&record.tags)
-                        .bind(&record.experimentcard_uid)
-                        .bind(&record.auditcard_uid)
-                        .bind(&record.pre_tag)
-                        .bind(&record.build_tag)
-                        .bind(&record.username)
-                        .bind(&record.opsml_version)
-                        .bind(&record.uid)
-                        .execute(&self.pool)
-                        .await?;
-                    Ok(())
-                }
-                _ => {
-                    return Err(SqlError::InvalidCardType);
-                }
-            },
-            CardTable::Experiment => match card {
-                ServerCard::Experiment(record) => {
-                    let query = SqliteQueryHelper::get_experimentcard_update_query();
-                    sqlx::query(&query)
-                        .bind(&record.app_env)
-                        .bind(&record.name)
-                        .bind(&record.space)
-                        .bind(record.major)
-                        .bind(record.minor)
-                        .bind(record.patch)
-                        .bind(&record.version)
-                        .bind(&record.tags)
-                        .bind(&record.datacard_uids)
-                        .bind(&record.modelcard_uids)
-                        .bind(&record.promptcard_uids)
-                        .bind(&record.service_card_uids)
-                        .bind(&record.experimentcard_uids)
-                        .bind(&record.pre_tag)
-                        .bind(&record.build_tag)
-                        .bind(&record.username)
-                        .bind(&record.opsml_version)
-                        .bind(&record.uid)
-                        .execute(&self.pool)
-                        .await?;
-                    Ok(())
-                }
-                _ => {
-                    return Err(SqlError::InvalidCardType);
-                }
-            },
-            CardTable::Audit => match card {
-                ServerCard::Audit(record) => {
-                    let query = SqliteQueryHelper::get_auditcard_update_query();
-                    sqlx::query(&query)
-                        .bind(&record.app_env)
-                        .bind(&record.name)
-                        .bind(&record.space)
-                        .bind(record.major)
-                        .bind(record.minor)
-                        .bind(record.patch)
-                        .bind(&record.version)
-                        .bind(&record.tags)
-                        .bind(record.approved)
-                        .bind(&record.datacard_uids)
-                        .bind(&record.modelcard_uids)
-                        .bind(&record.experimentcard_uids)
-                        .bind(&record.pre_tag)
-                        .bind(&record.build_tag)
-                        .bind(&record.username)
-                        .bind(&record.opsml_version)
-                        .bind(&record.uid)
-                        .execute(&self.pool)
-                        .await?;
-                    Ok(())
-                }
-                _ => {
-                    return Err(SqlError::InvalidCardType);
-                }
-            },
-
-            CardTable::Prompt => match card {
-                ServerCard::Prompt(record) => {
-                    let query = SqliteQueryHelper::get_promptcard_update_query();
-                    sqlx::query(&query)
-                        .bind(&record.app_env)
-                        .bind(&record.name)
-                        .bind(&record.space)
-                        .bind(record.major)
-                        .bind(record.minor)
-                        .bind(record.patch)
-                        .bind(&record.version)
-                        .bind(&record.tags)
-                        .bind(&record.experimentcard_uid)
-                        .bind(&record.auditcard_uid)
-                        .bind(&record.pre_tag)
-                        .bind(&record.build_tag)
-                        .bind(&record.username)
-                        .bind(&record.opsml_version)
-                        .bind(&record.uid)
-                        .execute(&self.pool)
-                        .await?;
-                    Ok(())
-                }
-
-                _ => {
-                    return Err(SqlError::InvalidCardType);
-                }
-            },
-
-            CardTable::Service => match card {
-                ServerCard::Service(record) => {
-                    let query = SqliteQueryHelper::get_servicecard_update_query();
-                    sqlx::query(&query)
-                        .bind(&record.app_env)
-                        .bind(&record.name)
-                        .bind(&record.space)
-                        .bind(record.major)
-                        .bind(record.minor)
-                        .bind(record.patch)
-                        .bind(&record.version)
-                        .bind(&record.cards)
-                        .bind(&record.username)
-                        .bind(&record.opsml_version)
-                        .bind(&record.uid)
-                        .execute(&self.pool)
-                        .await?;
-                    Ok(())
-                }
-                _ => {
-                    return Err(SqlError::InvalidCardType);
-                }
-            },
-
-            _ => {
-                return Err(SqlError::InvalidTableName);
-            }
-        }
-    }
-
-    /// Get unique space names
-    ///
-    /// # Arguments
-    ///
-    /// * `table` - The table to query
-    ///
-    /// # Returns
-    ///
-    /// * `Vec<String>` - A vector of unique space names
-    async fn get_unique_space_names(&self, table: &CardTable) -> Result<Vec<String>, SqlError> {
-        let query = format!("SELECT DISTINCT space FROM {table}");
-        let repos: Vec<String> = sqlx::query_scalar(&query).fetch_all(&self.pool).await?;
-
-        Ok(repos)
-    }
-
-    /// Query stats for a table
-    ///
-    /// # Arguments
-    ///
-    /// * `table` - The table to query
-    /// * `search_term` - The search term to query
-    ///
-    /// # Returns
-    ///
-    /// * `HashMap<String, i32>` - A hashmap of the stats
-    ///
-    async fn query_stats(
-        &self,
-        table: &CardTable,
-        search_term: Option<&str>,
-        space: Option<&str>,
-    ) -> Result<QueryStats, SqlError> {
-        let query = SqliteQueryHelper::get_query_stats_query(table);
-
-        // if search_term is not None, format with %search_term%, else None
-        let stats: QueryStats = sqlx::query_as(&query)
-            .bind(search_term.map(|term| format!("%{term}%")))
-            .bind(space)
-            .fetch_one(&self.pool)
-            .await?;
-
-        Ok(stats)
-    }
-
-    /// Query a page of cards
-    ///
-    /// # Arguments
-    ///
-    /// * `sort_by` - The field to sort by
-    /// * `page` - The page number
-    /// * `search_term` - The search term to query
-    /// * `space` - The space to query
-    /// * `table` - The table to query
-    ///
-    /// # Returns
-    ///
-    /// * `Vec<CardSummary>` - A vector of card summaries
-    async fn query_page(
-        &self,
-        sort_by: &str,
-        page: i32,
-        search_term: Option<&str>,
-        space: Option<&str>,
-        table: &CardTable,
-    ) -> Result<Vec<CardSummary>, SqlError> {
-        let query = SqliteQueryHelper::get_query_page_query(table, sort_by);
-
-        let lower_bound = (page * 30) - 30;
-        let upper_bound = page * 30;
-
-        let records: Vec<CardSummary> = sqlx::query_as(&query)
-            .bind(space)
-            .bind(search_term)
-            .bind(search_term.map(|term| format!("%{term}%")))
-            .bind(lower_bound)
-            .bind(upper_bound)
-            .fetch_all(&self.pool)
-            .await?;
-
-        Ok(records)
-    }
-
-    async fn version_page(
-        &self,
-        page: i32,
-        space: Option<&str>,
-        name: Option<&str>,
-        table: &CardTable,
-    ) -> Result<Vec<VersionSummary>, SqlError> {
-        let query = SqliteQueryHelper::get_version_page_query(table);
-
-        let lower_bound = (page * 30) - 30;
-        let upper_bound = page * 30;
-
-        let records: Vec<VersionSummary> = sqlx::query_as(&query)
-            .bind(space)
-            .bind(name)
-            .bind(lower_bound)
-            .bind(upper_bound)
-            .fetch_all(&self.pool)
-            .await?;
-
-        Ok(records)
-    }
-
-    #[instrument(skip_all)]
-    async fn delete_card(
-        &self,
-        table: &CardTable,
-        uid: &str,
-    ) -> Result<(String, String), SqlError> {
-        // SQLite doesn't support RETURNING clause, so we need to do this in two steps
-        debug!("Deleting card with uid: {uid} from table: {table}");
-        let select_query = format!("SELECT space, name FROM {table} WHERE uid = ?");
-        let (space, name): (String, String) = sqlx::query_as(&select_query)
-            .bind(uid)
-            .fetch_one(&self.pool)
-            .await?;
-
-        let delete_query = format!("DELETE FROM {table} WHERE uid = ?");
-        sqlx::query(&delete_query)
-            .bind(uid)
-            .execute(&self.pool)
-            .await?;
-
-        Ok((space, name))
-    }
-
-    async fn insert_experiment_metric(&self, record: &MetricRecord) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_experiment_metric_insert_query();
-
-        sqlx::query(&query)
-            .bind(&record.experiment_uid)
-            .bind(&record.name)
-            .bind(record.value)
-            .bind(record.step)
-            .bind(record.timestamp)
-            .bind(record.is_eval)
-            .execute(&self.pool)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn insert_experiment_metrics<'life1>(
-        &self,
-        records: &'life1 [MetricRecord],
-    ) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_experiment_metrics_insert_query(records.len());
-
-        let mut query_builder = sqlx::query(&query);
-
-        for r in records {
-            query_builder = query_builder
-                .bind(&r.experiment_uid)
-                .bind(&r.name)
-                .bind(r.value)
-                .bind(r.step)
-                .bind(r.timestamp)
-                .bind(r.is_eval);
-        }
-
-        query_builder.execute(&self.pool).await?;
-
-        Ok(())
-    }
-
-    async fn get_experiment_metric<'life2>(
-        &self,
-        uid: &str,
-        names: &'life2 [String],
-        is_eval: Option<bool>,
-    ) -> Result<Vec<MetricRecord>, SqlError> {
-        let (query, bindings) = SqliteQueryHelper::get_experiment_metric_query(names, is_eval);
-        let mut query_builder = sqlx::query_as::<sqlx::Sqlite, MetricRecord>(&query).bind(uid);
-
-        for binding in bindings {
-            query_builder = query_builder.bind(binding);
-        }
-
-        let records: Vec<MetricRecord> = query_builder.fetch_all(&self.pool).await?;
-
-        Ok(records)
-    }
-
-    async fn get_experiment_metric_names(&self, uid: &str) -> Result<Vec<String>, SqlError> {
-        let query = format!(
-            "SELECT DISTINCT name FROM {} WHERE experiment_uid = ?1",
-            CardTable::Metrics
-        );
-
-        let records: Vec<String> = sqlx::query_scalar(&query)
-            .bind(uid)
-            .fetch_all(&self.pool)
-            .await?;
-
-        Ok(records)
-    }
-
-    async fn insert_hardware_metrics(
-        &self,
-        record: &HardwareMetricsRecord,
-    ) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_hardware_metrics_insert_query();
-        sqlx::query(&query)
-            .bind(&record.experiment_uid)
-            .bind(record.created_at)
-            .bind(record.cpu_percent_utilization)
-            .bind(&record.cpu_percent_per_core)
-            .bind(record.free_memory)
-            .bind(record.total_memory)
-            .bind(record.used_memory)
-            .bind(record.available_memory)
-            .bind(record.used_percent_memory)
-            .bind(record.bytes_recv)
-            .bind(record.bytes_sent)
-            .execute(&self.pool)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn get_hardware_metric(&self, uid: &str) -> Result<Vec<HardwareMetricsRecord>, SqlError> {
-        let query = SqliteQueryHelper::get_hardware_metric_query();
-
-        let records: Vec<HardwareMetricsRecord> = sqlx::query_as(&query)
-            .bind(uid)
-            .fetch_all(&self.pool)
-            .await?;
-
-        Ok(records)
-    }
-
-    async fn insert_experiment_parameters<'life1>(
-        &self,
-        records: &'life1 [ParameterRecord],
-    ) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_experiment_parameters_insert_query(records.len());
-
-        let mut query_builder = sqlx::query(&query);
-
-        for record in records {
-            query_builder = query_builder
-                .bind(&record.experiment_uid)
-                .bind(&record.name)
-                .bind(&record.value);
-        }
-
-        query_builder.execute(&self.pool).await?;
-
-        Ok(())
-    }
-
-    async fn get_experiment_parameter<'life2>(
-        &self,
-        uid: &str,
-        names: &'life2 [String],
-    ) -> Result<Vec<ParameterRecord>, SqlError> {
-        let (query, bindings) = SqliteQueryHelper::get_experiment_parameter_query(names);
-        let mut query_builder = sqlx::query_as::<_, ParameterRecord>(&query).bind(uid);
-
-        for binding in bindings {
-            query_builder = query_builder.bind(binding);
-        }
-
-        let records: Vec<ParameterRecord> = query_builder.fetch_all(&self.pool).await?;
-
-        Ok(records)
-    }
-
-    async fn insert_user(&self, user: &User) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_user_insert_query();
-
-        let hashed_recovery_codes = serde_json::to_string(&user.hashed_recovery_codes)?;
-        let group_permissions = serde_json::to_string(&user.group_permissions)?;
-        let permissions = serde_json::to_string(&user.permissions)?;
-        let favorite_spaces = serde_json::to_string(&user.favorite_spaces)?;
-
-        sqlx::query(&query)
-            .bind(&user.username)
-            .bind(&user.password_hash)
-            .bind(&hashed_recovery_codes)
-            .bind(&permissions)
-            .bind(&group_permissions)
-            .bind(&favorite_spaces)
-            .bind(&user.role)
-            .bind(user.active)
-            .bind(&user.email)
-            .bind(&user.authentication_type)
-            .execute(&self.pool)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn get_user(
-        &self,
-        username: &str,
-        auth_type: Option<&str>,
-    ) -> Result<Option<User>, SqlError> {
-        let query = match auth_type {
-            Some(_) => SqliteQueryHelper::get_user_query_by_auth_type(),
-            None => SqliteQueryHelper::get_user_query(),
-        };
-
-        let mut query_builder = sqlx::query_as(&query).bind(username);
-
-        if let Some(auth_type) = auth_type {
-            query_builder = query_builder.bind(auth_type);
-        }
-
-        let user: Option<User> = query_builder.fetch_optional(&self.pool).await?;
-
-        Ok(user)
-    }
-
-    async fn get_users(&self) -> Result<Vec<User>, SqlError> {
-        let query = SqliteQueryHelper::get_users_query();
-
-        let users = sqlx::query_as::<_, User>(&query)
-            .fetch_all(&self.pool)
-            .await?;
-
-        Ok(users)
-    }
-
-    async fn is_last_admin(&self, username: &str) -> Result<bool, SqlError> {
-        // Count admins in the system
-        let query = SqliteQueryHelper::get_last_admin_query();
-
-        let admins: Vec<String> = sqlx::query_scalar(&query).fetch_all(&self.pool).await?;
-
-        // If there are no other admins, this is the last one
-        if admins.len() > 1 {
-            return Ok(false);
-        }
-
-        // no admins found
-        if admins.is_empty() {
-            return Ok(false);
-        }
-
-        // check if the username is the last admin
-        Ok(admins.len() == 1 && admins[0] == username)
-    }
-
-    async fn delete_user(&self, username: &str) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_user_delete_query();
-
-        sqlx::query(&query)
-            .bind(username)
-            .execute(&self.pool)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn update_user(&self, user: &User) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_user_update_query();
-
-        let hashed_recovery_codes = serde_json::to_string(&user.hashed_recovery_codes)?;
-        let group_permissions = serde_json::to_string(&user.group_permissions)?;
-        let permissions = serde_json::to_string(&user.permissions)?;
-        let favorite_spaces = serde_json::to_string(&user.favorite_spaces)?;
-
-        sqlx::query(&query)
-            .bind(user.active)
-            .bind(&user.password_hash)
-            .bind(&hashed_recovery_codes)
-            .bind(&permissions)
-            .bind(&group_permissions)
-            .bind(&favorite_spaces)
-            .bind(&user.refresh_token)
-            .bind(&user.email)
-            .bind(&user.authentication_type)
-            .bind(&user.username)
-            .bind(&user.authentication_type)
-            .execute(&self.pool)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn insert_artifact_key(&self, key: &ArtifactKey) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_artifact_key_insert_query();
-        sqlx::query(&query)
-            .bind(&key.uid)
-            .bind(&key.space)
-            .bind(key.registry_type.to_string())
-            .bind(key.encrypted_key.clone())
-            .bind(&key.storage_key)
-            .execute(&self.pool)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn get_artifact_key(
-        &self,
-        uid: &str,
-        registry_type: &str,
-    ) -> Result<ArtifactKey, SqlError> {
-        let query = SqliteQueryHelper::get_artifact_key_select_query();
-
-        let key: (String, String, String, Vec<u8>, String) = sqlx::query_as(&query)
-            .bind(uid)
-            .bind(registry_type)
-            .fetch_one(&self.pool)
-            .await?;
-
-        Ok(ArtifactKey {
-            uid: key.0,
-            space: key.1,
-            registry_type: RegistryType::from_string(&key.2)?,
-            encrypted_key: key.3,
-            storage_key: key.4,
-        })
-    }
-
-    async fn update_artifact_key(&self, key: &ArtifactKey) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_artifact_key_update_query();
-        sqlx::query(&query)
-            .bind(key.encrypted_key.clone())
-            .bind(&key.uid)
-            .bind(key.registry_type.to_string())
-            .execute(&self.pool)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn insert_audit_event(&self, event: AuditEvent) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_audit_event_insert_query();
-        sqlx::query(&query)
-            .bind(event.username)
-            .bind(event.client_ip)
-            .bind(event.user_agent)
-            .bind(event.operation.to_string())
-            .bind(event.resource_type.to_string())
-            .bind(event.resource_id)
-            .bind(event.access_location)
-            .bind(event.status.to_string())
-            .bind(event.error_message)
-            .bind(event.metadata)
-            .bind(event.registry_type.map(|r| r.to_string()))
-            .bind(event.route)
-            .execute(&self.pool)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn get_card_key_for_loading(
-        &self,
-        table: &CardTable,
-        query_args: &CardQueryArgs,
-    ) -> Result<ArtifactKey, SqlError> {
-        let query = SqliteQueryHelper::get_load_card_query(table, query_args)?;
-
-        let key: (String, String, String, Vec<u8>, String) = sqlx::query_as(&query)
-            .bind(query_args.uid.as_ref())
-            .bind(query_args.space.as_ref())
-            .bind(query_args.name.as_ref())
-            .bind(query_args.max_date.as_ref())
-            .bind(query_args.limit.unwrap_or(1))
-            .fetch_one(&self.pool)
-            .await?;
-
-        Ok(ArtifactKey {
-            uid: key.0,
-            space: key.1,
-            registry_type: RegistryType::from_string(&key.2)?,
-            encrypted_key: key.3,
-            storage_key: key.4,
-        })
-    }
-
-    async fn get_artifact_key_from_path(
-        &self,
-        storage_path: &str,
-        registry_type: &str,
-    ) -> Result<Option<ArtifactKey>, SqlError> {
-        let query = SqliteQueryHelper::get_artifact_key_from_storage_path_query();
-
-        let key: Option<(String, String, String, Vec<u8>, String)> = sqlx::query_as(&query)
-            .bind(storage_path)
-            .bind(registry_type)
-            .fetch_optional(&self.pool)
-            .await?;
-
-        return match key {
-            Some(k) => Ok(Some(ArtifactKey {
-                uid: k.0,
-                space: k.1,
-                registry_type: RegistryType::from_string(&k.2)?,
-                encrypted_key: k.3,
-                storage_key: k.4,
-            })),
-            None => Ok(None),
-        };
-    }
-
-    async fn delete_artifact_key(&self, uid: &str, registry_type: &str) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_artifact_key_delete_query();
-        sqlx::query(&query)
-            .bind(uid)
-            .bind(registry_type)
-            .execute(&self.pool)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn insert_space_record(&self, space: &SpaceRecord) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_insert_space_record_query();
-        sqlx::query(&query)
-            .bind(&space.space)
-            .bind(&space.description)
-            .execute(&self.pool)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn insert_space_name_record(&self, event: &SpaceNameEvent) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_insert_space_name_record_query();
-        sqlx::query(&query)
-            .bind(&event.space)
-            .bind(&event.name)
-            .bind(event.registry_type.to_string())
-            .execute(&self.pool)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn get_all_space_stats(&self) -> Result<Vec<SpaceStats>, SqlError> {
-        let query = SqliteQueryHelper::get_all_space_stats_query();
-        let spaces: Vec<SqlSpaceRecord> = sqlx::query_as(&query).fetch_all(&self.pool).await?;
-
-        Ok(spaces
-            .into_iter()
-            .map(|s| SpaceStats {
-                space: s.0,
-                model_count: s.1,
-                data_count: s.2,
-                prompt_count: s.3,
-                experiment_count: s.4,
-            })
-            .collect())
-    }
-
-    async fn get_space_record(&self, space: &str) -> Result<Option<SpaceRecord>, SqlError> {
-        let query = SqliteQueryHelper::get_space_record_query();
-        let record: Option<(String, String)> = sqlx::query_as(&query)
-            .bind(space)
-            .fetch_optional(&self.pool)
-            .await?;
-
-        Ok(record.map(|r| SpaceRecord {
-            space: r.0,
-            description: r.1,
-        }))
-    }
-
-    async fn update_space_record(&self, space: &SpaceRecord) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_update_space_record_query();
-        sqlx::query(&query)
-            .bind(&space.description)
-            .bind(&space.space)
-            .execute(&self.pool)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn delete_space_record(&self, space: &str) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_delete_space_record_query();
-        sqlx::query(&query).bind(space).execute(&self.pool).await?;
-
-        Ok(())
-    }
-
-    async fn delete_space_name_record(
-        &self,
-        space: &str,
-        name: &str,
-        registry_type: &RegistryType,
-    ) -> Result<(), SqlError> {
-        let query = SqliteQueryHelper::get_delete_space_name_record_query();
-        sqlx::query(&query)
-            .bind(space)
-            .bind(name)
-            .bind(registry_type.to_string())
-            .execute(&self.pool)
-            .await?;
-
-        Ok(())
-    }
+pub trait BaseSqliteClient {
+    fn pool(&self) -> &Pool<Sqlite>;
 }
 
 #[cfg(test)]
 mod tests {
 
-    use crate::schemas::ServiceCardRecord;
-
     use super::*;
 
-    use opsml_types::contracts::ArtifactType;
-    use opsml_types::{contracts::SpaceNameEvent, RegistryType, SqlType};
+    use crate::schemas::schema::{
+        ArtifactSqlRecord, AuditCardRecord, CardResults, DataCardRecord, ExperimentCardRecord,
+        HardwareMetricsRecord, MetricRecord, ModelCardRecord, ParameterRecord, PromptCardRecord,
+        ServerCard, ServiceCardRecord, User,
+    };
+    use crate::traits::{
+        ArtifactLogicTrait, AuditLogicTrait, CardLogicTrait, ExperimentLogicTrait, SpaceLogicTrait,
+        UserLogicTrait,
+    };
+    use opsml_settings::config::DatabaseSettings;
+    use opsml_types::contracts::{ArtifactKey, ArtifactQueryArgs, AuditEvent, SpaceNameEvent};
+    use opsml_types::SqlType;
+    use opsml_types::{
+        cards::CardTable,
+        contracts::{ArtifactType, CardQueryArgs},
+        RegistryType,
+    };
     use opsml_utils::utils::get_utc_datetime;
+    use semver::Version;
     use std::env;
 
     const SPACE: &str = "space";
@@ -1283,14 +143,14 @@ mod tests {
         };
 
         // Test Insert
-        client.insert_card(table, &card).await?;
+        client.card.insert_card(table, &card).await?;
 
         // Verify Insert
         let card_args = CardQueryArgs {
             uid: Some(uid.clone()),
             ..Default::default()
         };
-        let results = client.query_cards(table, &card_args).await?;
+        let results = client.card.query_cards(table, &card_args).await?;
         assert_eq!(results.len(), 1);
 
         // Create updated card with new name
@@ -1349,10 +209,10 @@ mod tests {
         };
 
         // Test Update
-        client.update_card(table, &updated_card).await?;
+        client.card.update_card(table, &updated_card).await?;
 
         // Verify Update
-        let updated_results = client.query_cards(table, &card_args).await?;
+        let updated_results = client.card.query_cards(table, &card_args).await?;
         assert_eq!(updated_results.len(), 1);
 
         // Verify updated name
@@ -1366,10 +226,10 @@ mod tests {
         }
 
         // delete card
-        client.delete_card(table, &uid).await?;
+        client.card.delete_card(table, &uid).await?;
 
         // Verify Delete
-        let deleted_results = client.query_cards(table, &card_args).await?;
+        let deleted_results = client.card.query_cards(table, &card_args).await?;
         assert_eq!(deleted_results.len(), 0);
 
         Ok(())
@@ -1425,6 +285,7 @@ mod tests {
         // query all versions
         // get versions (should return 1)
         let versions = client
+            .card
             .get_versions(&CardTable::Data, "repo1", "Data1", None)
             .await
             .unwrap();
@@ -1432,18 +293,21 @@ mod tests {
 
         // check star pattern
         let versions = client
+            .card
             .get_versions(&CardTable::Data, "repo1", "Data1", Some("*".to_string()))
             .await
             .unwrap();
         assert_eq!(versions.len(), 10);
 
         let versions = client
+            .card
             .get_versions(&CardTable::Data, "repo1", "Data1", Some("1.*".to_string()))
             .await
             .unwrap();
         assert_eq!(versions.len(), 4);
 
         let versions = client
+            .card
             .get_versions(
                 &CardTable::Data,
                 "repo1",
@@ -1456,6 +320,7 @@ mod tests {
 
         // check tilde pattern
         let versions = client
+            .card
             .get_versions(&CardTable::Data, "repo1", "Data1", Some("~1".to_string()))
             .await
             .unwrap();
@@ -1463,6 +328,7 @@ mod tests {
 
         // check tilde pattern
         let versions = client
+            .card
             .get_versions(&CardTable::Data, "repo1", "Data1", Some("~1.1".to_string()))
             .await
             .unwrap();
@@ -1470,6 +336,7 @@ mod tests {
 
         // check tilde pattern
         let versions = client
+            .card
             .get_versions(
                 &CardTable::Data,
                 "repo1",
@@ -1481,6 +348,7 @@ mod tests {
         assert_eq!(versions.len(), 1);
 
         let versions = client
+            .card
             .get_versions(
                 &CardTable::Data,
                 "repo1",
@@ -1492,6 +360,7 @@ mod tests {
         assert_eq!(versions.len(), 2);
 
         let versions = client
+            .card
             .get_versions(
                 &CardTable::Data,
                 "repo1",
@@ -1503,6 +372,7 @@ mod tests {
         assert_eq!(versions.len(), 1);
 
         let versions = client
+            .card
             .get_versions(&CardTable::Data, "repo1", "Data1", Some("2".to_string()))
             .await
             .unwrap();
@@ -1564,6 +434,7 @@ mod tests {
 
         // get unique space names
         let repos = client
+            .card
             .get_unique_space_names(&CardTable::Model)
             .await
             .unwrap();
@@ -1591,6 +462,7 @@ mod tests {
 
         // query stats
         let stats = client
+            .card
             .query_stats(&CardTable::Model, None, None)
             .await
             .unwrap();
@@ -1601,6 +473,7 @@ mod tests {
 
         // query stats with search term
         let stats = client
+            .card
             .query_stats(&CardTable::Model, Some("Model1"), None)
             .await
             .unwrap();
@@ -1608,6 +481,7 @@ mod tests {
         assert_eq!(stats.nbr_names, 2); // for Model1 and Model10
 
         let stats = client
+            .card
             .query_stats(&CardTable::Model, Some("Model1"), Some("repo1"))
             .await
             .unwrap();
@@ -1635,6 +509,7 @@ mod tests {
 
         // query page
         let results = client
+            .card
             .query_page("name", 1, None, None, &CardTable::Data)
             .await
             .unwrap();
@@ -1643,6 +518,7 @@ mod tests {
 
         // query page
         let results = client
+            .card
             .query_page("name", 1, None, None, &CardTable::Model)
             .await
             .unwrap();
@@ -1651,6 +527,7 @@ mod tests {
 
         // query page
         let results = client
+            .card
             .query_page("name", 1, None, Some("repo3"), &CardTable::Model)
             .await
             .unwrap();
@@ -1678,6 +555,7 @@ mod tests {
 
         // query page
         let results = client
+            .card
             .version_page(1, Some("repo1"), Some("Model1"), &CardTable::Model)
             .await
             .unwrap();
@@ -1717,7 +595,7 @@ mod tests {
                 ..Default::default()
             };
 
-            client.insert_experiment_metric(&metric).await.unwrap();
+            client.exp.insert_experiment_metric(&metric).await.unwrap();
         }
 
         for name in eval_metric_names {
@@ -1729,19 +607,21 @@ mod tests {
                 ..Default::default()
             };
 
-            client.insert_experiment_metric(&metric).await.unwrap();
+            client.exp.insert_experiment_metric(&metric).await.unwrap();
         }
 
         let records = client
+            .exp
             .get_experiment_metric(&uid, &Vec::new(), None)
             .await
             .unwrap();
-        let names = client.get_experiment_metric_names(&uid).await.unwrap();
+        let names = client.exp.get_experiment_metric_names(&uid).await.unwrap();
 
         assert_eq!(records.len(), 5);
         assert_eq!(names.len(), 5);
 
         let eval_records = client
+            .exp
             .get_experiment_metric(&uid, &Vec::new(), Some(true))
             .await
             .unwrap();
@@ -1764,9 +644,14 @@ mod tests {
             },
         ];
 
-        client.insert_experiment_metrics(&records).await.unwrap();
+        client
+            .exp
+            .insert_experiment_metrics(&records)
+            .await
+            .unwrap();
 
         let records = client
+            .exp
             .get_experiment_metric(&uid, &Vec::new(), None)
             .await
             .unwrap();
@@ -1803,8 +688,8 @@ mod tests {
             ..Default::default()
         };
 
-        client.insert_hardware_metrics(&metric).await.unwrap();
-        let records = client.get_hardware_metric(&uid).await.unwrap();
+        client.exp.insert_hardware_metrics(&metric).await.unwrap();
+        let records = client.exp.get_hardware_metric(&uid).await.unwrap();
 
         assert_eq!(records.len(), 1);
 
@@ -1842,8 +727,13 @@ mod tests {
             params.push(parameter.clone());
         }
 
-        client.insert_experiment_parameters(&params).await.unwrap();
+        client
+            .exp
+            .insert_experiment_parameters(&params)
+            .await
+            .unwrap();
         let records = client
+            .exp
             .get_experiment_parameter(&uid, &Vec::new())
             .await
             .unwrap();
@@ -1851,6 +741,7 @@ mod tests {
         assert_eq!(records.len(), 10);
 
         let param_records = client
+            .exp
             .get_experiment_parameter(&uid, &["param1".to_string()])
             .await
             .unwrap();
@@ -1888,10 +779,10 @@ mod tests {
 
         let sso_user = User::new_from_sso("sso_user", "user@email.com");
 
-        client.insert_user(&user).await.unwrap();
-        client.insert_user(&sso_user).await.unwrap();
+        client.user.insert_user(&user).await.unwrap();
+        client.user.insert_user(&sso_user).await.unwrap();
 
-        let mut user_to_update = client.get_user("user", None).await.unwrap().unwrap();
+        let mut user_to_update = client.user.get_user("user", None).await.unwrap().unwrap();
 
         assert_eq!(user_to_update.username, "user");
         assert_eq!(user_to_update.password_hash, "pass");
@@ -1902,17 +793,18 @@ mod tests {
         user_to_update.active = false;
         user_to_update.refresh_token = Some("token".to_string());
 
-        client.update_user(&user_to_update).await.unwrap();
-        let user = client.get_user("user", None).await.unwrap().unwrap();
+        client.user.update_user(&user_to_update).await.unwrap();
+        let user = client.user.get_user("user", None).await.unwrap().unwrap();
 
         assert!(!user.active);
         assert_eq!(user.refresh_token.unwrap(), "token");
 
         // get users
-        let users = client.get_users().await.unwrap();
+        let users = client.user.get_users().await.unwrap();
         assert_eq!(users.len(), 2);
 
         let user = client
+            .user
             .get_user("sso_user", Some("sso"))
             .await
             .unwrap()
@@ -1920,14 +812,14 @@ mod tests {
         assert!(user.active);
 
         // delete
-        client.delete_user("sso_user").await.unwrap();
+        client.user.delete_user("sso_user").await.unwrap();
 
         // get last admin
-        let is_last_admin = client.is_last_admin("user").await.unwrap();
+        let is_last_admin = client.user.is_last_admin("user").await.unwrap();
         assert!(!is_last_admin);
 
         // delete
-        client.delete_user("user").await.unwrap();
+        client.user.delete_user("user").await.unwrap();
 
         cleanup();
     }
@@ -1954,9 +846,10 @@ mod tests {
             storage_key: "opsml_registry".to_string(),
         };
 
-        client.insert_artifact_key(&key).await.unwrap();
+        client.artifact.insert_artifact_key(&key).await.unwrap();
 
         let key = client
+            .artifact
             .get_artifact_key(&key.uid, &key.registry_type.to_string())
             .await
             .unwrap();
@@ -1973,9 +866,10 @@ mod tests {
             storage_key: "opsml_registry".to_string(),
         };
 
-        client.update_artifact_key(&key).await.unwrap();
+        client.artifact.update_artifact_key(&key).await.unwrap();
 
         let key = client
+            .artifact
             .get_artifact_key(&key.uid, &key.registry_type.to_string())
             .await
             .unwrap();
@@ -1996,6 +890,7 @@ mod tests {
         let client = SqliteClient::new(&config).await.unwrap();
 
         client
+            .audit
             .insert_audit_event(AuditEvent::default())
             .await
             .unwrap();
@@ -2024,7 +919,11 @@ mod tests {
         let data_card = DataCardRecord::default();
         let card = ServerCard::Data(data_card.clone());
 
-        client.insert_card(&CardTable::Data, &card).await.unwrap();
+        client
+            .card
+            .insert_card(&CardTable::Data, &card)
+            .await
+            .unwrap();
         let encrypted_key: Vec<u8> = (0..32).collect();
         let key = ArtifactKey {
             uid: data_card.uid.clone(),
@@ -2034,10 +933,11 @@ mod tests {
             storage_key: "opsml_registry".to_string(),
         };
 
-        client.insert_artifact_key(&key).await.unwrap();
+        client.artifact.insert_artifact_key(&key).await.unwrap();
 
         // test uid (testing to ensure it doesnt fail)
         let _key = client
+            .card
             .get_card_key_for_loading(
                 &CardTable::Data,
                 &CardQueryArgs {
@@ -2051,6 +951,7 @@ mod tests {
 
         // test args
         let key = client
+            .card
             .get_card_key_for_loading(
                 &CardTable::Data,
                 &CardQueryArgs {
@@ -2067,6 +968,7 @@ mod tests {
         assert_eq!(key.encrypted_key, encrypted_key);
 
         let _ = client
+            .artifact
             .get_artifact_key_from_path(&key.storage_key, &RegistryType::Data.to_string())
             .await
             .unwrap()
@@ -2074,6 +976,7 @@ mod tests {
 
         // delete
         client
+            .artifact
             .delete_artifact_key(&data_card.uid, &RegistryType::Data.to_string())
             .await
             .unwrap();
@@ -2094,29 +997,45 @@ mod tests {
         // insert datacard
         let data_card = DataCardRecord::default();
         let card = ServerCard::Data(data_card.clone());
-        client.insert_card(&CardTable::Data, &card).await.unwrap();
+        client
+            .card
+            .insert_card(&CardTable::Data, &card)
+            .await
+            .unwrap();
 
         // insert modelcard
         let model_card = ModelCardRecord::default();
         let card = ServerCard::Model(model_card.clone());
-        client.insert_card(&CardTable::Model, &card).await.unwrap();
+        client
+            .card
+            .insert_card(&CardTable::Model, &card)
+            .await
+            .unwrap();
 
         let space_event = SpaceNameEvent {
             space: data_card.space.clone(),
             name: data_card.name.clone(),
             registry_type: RegistryType::Data,
         };
-        client.insert_space_name_record(&space_event).await.unwrap();
+        client
+            .space
+            .insert_space_name_record(&space_event)
+            .await
+            .unwrap();
 
         let space_event = SpaceNameEvent {
             space: model_card.space.clone(),
             name: model_card.name.clone(),
             registry_type: RegistryType::Model,
         };
-        client.insert_space_name_record(&space_event).await.unwrap();
+        client
+            .space
+            .insert_space_name_record(&space_event)
+            .await
+            .unwrap();
 
         // get space stats
-        let stats = client.get_all_space_stats().await.unwrap();
+        let stats = client.space.get_all_space_stats().await.unwrap();
         assert_eq!(stats.len(), 1);
         // assert model_count
         assert_eq!(stats[0].data_count, 1);
@@ -2127,7 +1046,11 @@ mod tests {
             ..Default::default()
         };
         let card = ServerCard::Model(model_card2.clone());
-        client.insert_card(&CardTable::Model, &card).await.unwrap();
+        client
+            .card
+            .insert_card(&CardTable::Model, &card)
+            .await
+            .unwrap();
 
         // update space stats again
         let space_event = SpaceNameEvent {
@@ -2135,10 +1058,14 @@ mod tests {
             name: model_card2.name.clone(),
             registry_type: RegistryType::Model,
         };
-        client.insert_space_name_record(&space_event).await.unwrap();
+        client
+            .space
+            .insert_space_name_record(&space_event)
+            .await
+            .unwrap();
         // get space stats again
 
-        let stats = client.get_all_space_stats().await.unwrap();
+        let stats = client.space.get_all_space_stats().await.unwrap();
         assert_eq!(stats.len(), 1);
 
         // assert model_count
@@ -2146,6 +1073,7 @@ mod tests {
 
         // delete space name record
         client
+            .space
             .delete_space_name_record(&data_card.space, &data_card.name, &RegistryType::Data)
             .await
             .unwrap();
@@ -2174,6 +1102,7 @@ mod tests {
             ArtifactType::Generic.to_string(),
         );
         client
+            .artifact
             .insert_artifact_record(&artifact_record1)
             .await
             .unwrap();
@@ -2187,12 +1116,14 @@ mod tests {
         );
 
         client
+            .artifact
             .insert_artifact_record(&artifact_record2)
             .await
             .unwrap();
 
         // query artifacts
         let artifacts = client
+            .artifact
             .query_artifacts(&ArtifactQueryArgs {
                 space: Some(SPACE.to_string()),
                 name: Some(name.clone()),
