+++ crates/opsml_sql/src/sqlite/sql/mod.rs
@@ -1,7 +1,6 @@
 pub mod artifact;
 pub mod audit;
 pub mod card;
-pub mod client;
 pub mod experiment;
 pub mod space;
 pub mod user;
